Update branch to latest

+++ routes/admin.rb
@@ -219,6 +219,10 @@ post '/admin/config' do
     udv = params["user_defined_variables"].split(",")
     rat = params["report_assessment_types"].split(",")
 
+    if params["effort"]
+        config_options["effort"] = params["effort"].split(",")
+    end
+
     config_options["finding_types"] = ft
     config_options["user_defined_variables"] = udv
     config_options["port"] = params["port"]
@@ -295,7 +299,6 @@ post '/admin/plugins' do
     }
 
     @plugins.each do |plug|
-        p params
         if params[plug["name"]]
             plug["enabled"] = true
             File.open("./plugins/#{plug['name']}/plugin.json","w") do |f|
@@ -409,12 +412,16 @@ get '/admin/delete/templates/:id' do
 
     @xslt = Xslt.first(:id => params[:id])
 
-	if @xslt
-		@xslt.components.destroy
-		@xslt.destroy
-		File.delete(@xslt.xslt_location)
-		File.delete(@xslt.docx_location)
-	end
+  	if @xslt
+    		@xslt.components.destroy
+    		@xslt.destroy
+        if File.file?(@xslt.xslt_location)
+    		    File.delete(@xslt.xslt_location)
+        end
+        if File.file?(@xslt.docx_location)
+    		    File.delete(@xslt.docx_location)
+        end
+  	end
     redirect to('/admin/templates')
 end
 
@@ -425,23 +432,23 @@ post '/admin/templates/add' do
 
     @admin = true
 
-	xslt_file = "./templates/#{rand(36**36).to_s(36)}.xslt"
+    xslt_file = "./templates/#{rand(36**36).to_s(36)}.xslt"
 
     redirect to("/admin/templates/add") unless params[:file]
 
-	# reject if the file is above a certain limit
-	if params[:file][:tempfile].size > 100000000
-		return "File too large. 10MB limit"
-	end
+    # reject if the file is above a certain limit
+    if params[:file][:tempfile].size > 100000000
+        return "File too large. 10MB limit"
+    end
 
-	docx = "./templates/#{rand(36**36).to_s(36)}.docx"
-	File.open(docx, 'wb') {|f| f.write(params[:file][:tempfile].read) }
+    docx = "./templates/#{rand(36**36).to_s(36)}.docx"
+    File.open(docx, 'wb') {|f| f.write(params[:file][:tempfile].read) }
 
     error = false
     detail = ""
     begin
-	    xslt = generate_xslt(docx)
-		xslt_components = generate_xslt_components(docx)
+        xslt = generate_xslt(docx)
+        xslt_components = generate_xslt_components(docx)
     rescue ReportingError => detail
         error = true
     end
@@ -451,50 +458,52 @@ post '/admin/templates/add' do
         "The report template you uploaded threw an error when parsing:<p><p> #{detail.errorString}"
     else
 
-    	# open up a file handle and write the attachment
-	    File.open(xslt_file, 'wb') {|f| f.write(xslt) }
+      # open up a file handle and write the attachment
+      File.open(xslt_file, 'wb') {|f| f.write(xslt) }
+      #extract the screenshot names from the file
+      screenshot_names = xslt.scan(/\[!!(.*?)!!\]/)
+      # delete the file data from the attachment
+      datax = Hash.new
+      # to prevent traversal we hardcode this
+      datax["docx_location"] = "#{docx}"
+      datax["xslt_location"] = "#{xslt_file}"
+      datax["description"] = 	params[:description]
+      datax["report_type"] = params[:report_type]
+      datax["screenshot_names"] = screenshot_names.join(",")
+      data = url_escape_hash(datax)
+      data["finding_template"] = params[:finding_template] ? true : false
+      data["status_template"] = params[:status_template] ? true : false
 
-	    # delete the file data from the attachment
-	    datax = Hash.new
-	    # to prevent traversal we hardcode this
-	    datax["docx_location"] = "#{docx}"
-	    datax["xslt_location"] = "#{xslt_file}"
-	    datax["description"] = 	params[:description]
-	    datax["report_type"] = params[:report_type]
-	    data = url_escape_hash(datax)
-	    data["finding_template"] = params[:finding_template] ? true : false
-	    data["status_template"] = params[:status_template] ? true : false
+      @template = Xslt.first(:report_type => data["report_type"])
 
-	    @template = Xslt.first(:report_type => data["report_type"])
+      if @template
+          @template.update(:xslt_location => data["xslt_location"], :docx_location => data["docx_location"], :description => data["description"], :screenshot_names => data["screenshot_names"])
+	  @template.components.destroy
+      else
+          @template = Xslt.new(data)
+          @template.save
+      end
 
-	    if @template
-		    @template.update(:xslt_location => data["xslt_location"], :docx_location => data["docx_location"], :description => data["description"])
-			@template.components.destroy
-	    else
-		    @template = Xslt.new(data)
-		    @template.save
-	    end
+      # create a xslt file for each component
+      list_components_files = []
+      xslt_components.each do |component_name, component_xslt|
+          componentHash = Hash.new
+          componentHash['xslt_location'] = "./templates/#{rand(36**36).to_s(36)}.xslt"
+          componentHash['name'] = component_name
+          componentHash['xslt'] = @template
+          File.open(componentHash['xslt_location'], 'wb') {|f| f.write(component_xslt) }
+          list_components_files.push(componentHash)
+      end
 
-		# create a xslt file for each component
-		list_components_files = []
-		xslt_components.each do |component_name, component_xslt|
-			componentHash = Hash.new
-			componentHash['xslt_location'] = "./templates/#{rand(36**36).to_s(36)}.xslt"
-			componentHash['name'] = component_name
-			componentHash['xslt'] = @template
-			File.open(componentHash['xslt_location'], 'wb') {|f| f.write(component_xslt) }
-			list_components_files.push(componentHash)
-		end
+      # insert components into the db
+      list_components_files.each do |component|
+          @component = Xslt_component.new(component)
+          @component.save
+      end
+      redirect to("/admin/templates")
 
-		# insert components into the db
-		list_components_files.each do |component|
-			@component = Xslt_component.new(component)
-			@component.save
-		end
-	    redirect to("/admin/templates")
-
-        haml :add_template, :encode_html => true
-    end
+      haml :add_template, :encode_html => true
+  end
 end
 
 # Manage Templated Reports
@@ -541,7 +550,8 @@ post '/admin/templates/edit' do
 
     	# open up a file handle and write the attachment
 	    File.open(xslt_file, 'wb') {|f| f.write(xslt) }
-
+      #extract the screenshot names from the file
+      screenshot_names = xslt.scan(/\[!!(.*?)!!\]/)
 	    # delete the file data from the attachment
 	    datax = Hash.new
 	    # to prevent traversal we hardcode this
@@ -549,6 +559,7 @@ post '/admin/templates/edit' do
 	    datax["xslt_location"] = "#{xslt_file}"
 	    datax["description"] = 	params[:description]
 	    datax["report_type"] = params[:report_type]
+      datax["screenshot_names"] = screenshot_names.join(",")
 	    data = url_escape_hash(datax)
 	    data["finding_template"] = params[:finding_template] ? true : false
 	    data["status_template"] = params[:status_template] ? true : false
@@ -556,7 +567,7 @@ post '/admin/templates/edit' do
 	    @template = Xslt.first(:report_type => data["report_type"])
 
 	    if @template
-		    @template.update(:xslt_location => data["xslt_location"], :docx_location => data["docx_location"], :description => data["description"])
+        @template.update(:xslt_location => data["xslt_location"], :docx_location => data["docx_location"], :description => data["description"], :screenshot_names => data["screenshot_names"])
 			@template.components.destroy
 	    else
 		    @template = Xslt.new(data)
